refactor : nickname 변경 컴포넌트 분리
