+++ front/src/components/User/Info/index.tsx
@@ -1,12 +1,10 @@
 import React , { useRef, useCallback  } from 'react';
 import { useDispatch } from 'react-redux';
 import { MyInfo } from '../../../model/MyInfo';
-import { OPEN_ALERT } from '../../../actions/alert';
-import { UPDATE_NICKNAME_REQUEST , UPDATE_PROFILE_PIC_REQUEST } from '../../../actions/user';
+import { UPDATE_PROFILE_PIC_REQUEST } from '../../../actions/user';
 import { faPlus } from '@fortawesome/free-solid-svg-icons';
-import useValidation from '../../../hooks/useValidation';
-import Button from '../../../atoms/Buttons';
 import Icon from '../../../atoms/Icons';
+import NicknameChange from '../NicknameChange';
 import Logout from '../Logout';
 import Avatar from '../../Avatar';
 import Slot from '../../Slot';
@@ -21,11 +19,6 @@ const Info=({myInfo} : Props)=>{
     const dispatch = useDispatch();
     const imageInput = useRef<any>(null);
 
-    const [ nickname,  // 닉네임 
-            setNickname, 
-            nicknameError
-        ] = useValidation(myInfo.nickname,2,6);
-
     // 이미지 업로드 버튼 클릭시 
     const onUploadImage = useCallback(()=>{ 
         imageInput.current.click();
@@ -38,16 +31,6 @@ const Info=({myInfo} : Props)=>{
         dispatch({type:UPDATE_PROFILE_PIC_REQUEST, data:imageFormData}); // 아바타 변경 
     },[]);
 
-    // 닉네임 변경 
-    const onChangeNickname = useCallback((e)=>{ 
-        e.preventDefault();
-        dispatch({
-            type:UPDATE_NICKNAME_REQUEST,
-            data:{nickname:nickname},
-        })
-        dispatch({type:OPEN_ALERT, data:"닉네임이 변경되었습니다."}); // 확인 alert 
-    },[nickname]);
-
     return(
         <Container>
             <AvatarContainer>
@@ -58,23 +41,7 @@ const Info=({myInfo} : Props)=>{
                  </EditIcon>
                 <Avatar size={100}/>
             </AvatarContainer>
-            <Form>
-                <InputContainer>
-                    <Label>닉네임</Label>
-                    <Nickname
-                    value={nickname}
-                    onChange={setNickname}
-                    />
-                    {nicknameError && <ErrorMessage>닉네임은 2글자 이상 6글자 이하여야 합니다.</ErrorMessage> }
-                </InputContainer>
-                <ButtonContainer>
-                    <Button
-                    title="수정"
-                    onClick={onChangeNickname}
-                    disabled={nicknameError || nickname===myInfo.nickname}
-                    />
-                </ButtonContainer>
-            </Form>
+            <NicknameChange exNickname={myInfo.nickname}/>
             <ReviewCountContainer>
                 <Title>{myInfo.Reviews} 개의 리뷰를 쓰신 당신은!</Title>
                  <Slot reviewsCount={myInfo.Reviews}/>
@@ -136,46 +103,6 @@ const EditIcon = styled.div`
     &:hover{
         background-color:rgba(255,255,255,0.3);
     }
-`;
-
-const Form = styled.form`
-    display:flex;
-    align-items:center;
-    margin:auto;
-`;
-
-const InputContainer = styled.div`
-    width:100%;
-    display:flex;
-    flex-direction:column;
-    position:relative;
-`;
-
-const Label = styled.label`
-    font-size:1rem;
-    color:#cc00cc;
-`;
-
-const Nickname = styled.input.attrs({type:'text'})`
-    width:100%;
-    padding: 10px 15px;
-    font-size:1.2rem;
-    border:none;
-    border-bottom:3px solid #e6b3cc;
-
-    background-color:inherit;
-    color:inherit;
-`;
-
-const ErrorMessage = styled.div`
-    width:80%;
-    color:#ff3333;
-    position:absolute;
-    bottom:-40px;
-`;
-
-const ButtonContainer = styled.div`
-    flex-basis:40%;
 `;
 
 const Title = styled.div`
